// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Rol = {
  ADMIN: 'ADMIN',
  CAJERO: 'CAJERO'
} as const

export type Rol = (typeof Rol)[keyof typeof Rol]


export const CanalVenta = {
  MOSTRADOR: 'MOSTRADOR',
  WEB: 'WEB',
  MARKETPLACE: 'MARKETPLACE'
} as const

export type CanalVenta = (typeof CanalVenta)[keyof typeof CanalVenta]


export const EstadoVenta = {
  PENDIENTE: 'PENDIENTE',
  PAGADA: 'PAGADA',
  ENVIADA: 'ENVIADA',
  ENTREGADA: 'ENTREGADA',
  CANCELADA: 'CANCELADA'
} as const

export type EstadoVenta = (typeof EstadoVenta)[keyof typeof EstadoVenta]


export const MetodoPago = {
  EFECTIVO: 'EFECTIVO',
  TARJETA: 'TARJETA',
  TRANSFERENCIA: 'TRANSFERENCIA',
  CONTRA_ENTREGA: 'CONTRA_ENTREGA'
} as const

export type MetodoPago = (typeof MetodoPago)[keyof typeof MetodoPago]


export const EstadoPago = {
  PENDIENTE: 'PENDIENTE',
  CONFIRMADO: 'CONFIRMADO',
  RECHAZADO: 'RECHAZADO'
} as const

export type EstadoPago = (typeof EstadoPago)[keyof typeof EstadoPago]


export const TipoMovimiento = {
  ENTRADA: 'ENTRADA',
  VENTA: 'VENTA',
  AJUSTE: 'AJUSTE',
  DEVOLUCION: 'DEVOLUCION',
  CANCELACION: 'CANCELACION'
} as const

export type TipoMovimiento = (typeof TipoMovimiento)[keyof typeof TipoMovimiento]